resolve conversations
