+++ src/components/main/actionsPanel/navigatorPanel/helpers/helpers.ts
@@ -62,7 +62,9 @@ export const selectFirstNode = (
       }
     }
     if (firstNodeId !== "0") {
-      const _expandedItems = getExpandedItems(validNodeTree, [firstNodeId]);
+      const _expandedItems = getNeedToExpandNodeUids(validNodeTree, [
+        firstNodeId,
+      ]);
       dispatch(expandNodeTreeNodes(_expandedItems));
 
       console.log({ firstNodeId });
@@ -74,22 +76,6 @@ export const selectFirstNode = (
   return true;
 };
 
-export const getExpandedItems = (
-  validNodeTree: TNodeTreeData,
-  selectedNodeUids: TNodeUid[],
-): TNodeUid[] => {
-  const _expandedItems: TNodeUid[] = [];
-  let node = validNodeTree[selectedNodeUids[0]];
-  if (!node) {
-    return _expandedItems;
-  }
-  while (node.uid !== RootNodeUid) {
-    _expandedItems.push(node.uid);
-    node = validNodeTree[node.parentUid as TNodeUid];
-  }
-  _expandedItems.shift();
-  return _expandedItems;
-};
 export const setWorkspaceFavicon = (
   validNodeTree: TNodeTreeData,
   project: TProject,
