+++ src/components/main/stageView/iFrame/hooks/useMouseEvents.ts
@@ -3,16 +3,11 @@ import { useDispatch } from "react-redux";
 import { StageNodeIdAttr } from "@_node/file";
 import { getValidNodeUids } from "@_node/helpers";
 import { TNodeTreeData, TNodeUid } from "@_node/types";
-import {
-  setAppendExpandedNodeTreeNodes,
-  setExpandedNodeTreeNodes,
-  setHoveredNodeUid,
-} from "@_redux/main/nodeTree";
+import { setHoveredNodeUid } from "@_redux/main/nodeTree";
 import { setSelectedNodeUids } from "@_redux/main/nodeTree/event";
 import { getValidElementWithUid, selectAllText } from "../helpers";
 import { THtmlNodeData } from "@_node/node";
 import { setActivePanel } from "@_redux/main/processor";
-import { getExpandedItems } from "@_components/main/actionsPanel/navigatorPanel/helpers";
 
 interface IUseMouseEventsProps {
   contentRef: HTMLIFrameElement | null;
@@ -59,8 +54,6 @@ export const useMouseEvents = ({
           uids = validUids;
         }
         dispatch(setSelectedNodeUids(uids));
-        const expandedItems = getExpandedItems(nodeTreeRef.current, uids);
-        dispatch(setAppendExpandedNodeTreeNodes(expandedItems));
       }
 
       if (contentEditableUidRef.current && contentRef) {
